requirements.txt
Django==4.2.11
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1